[now-node] Migrate to TypeScript
This moves `@now/node` to being implemented in TypeScript.

It also eliminates the need for the `@now/node-server` builder, by
introducing a convention based on what was exported in the user code:

 * If the user code exports a function, then it's an http handler
   (current `@now/node` behavior).
 * If the user code exports an `http.Server`, then invoke `listen()` on
   an ephemeral port.
 * If the user code did not export a function or `http.Server`, then
   assume the user is invoking `.listen()` manually (old
   `@now/node-server` behavior).

Finally, this removes the port 3000 hard-coding that was previously
in-place, which was a problem for `now dev`. Now the `http.Server`
instance listens on an ephemeral port which is detected by the Now node
runtime.