Allowing main functions through the Go builder
This is a concept idea that should fix #252

This PR will replace any existing `main` function with a function called
`Main` with an apendix to the name composed of random numbers, as in:
`Main3831`. This will be done only if the `src` file has a `main`
package.  After that, knowing that a main function was present, this PR
makes sure that this `Main` function gets called within the `main`
function we declare in the build, right before
`now.Start(http.HandlerFunc(__NOW_HANDLER_FUNC_NAME))`.

If this PR is not ok, feel free to close it! I hope it serves at least
as a form of documentation for any further fixes to #252

Best regards to the Zeit team,
Daniel.