[now-build-utils] Add support for symlinks to `download()`
`now dev` is using this function to extract the `prepareCache()`
results, but upon extraction symlinks are lost and turned into
regular files, which breaks relative requires in Node.js modules.

This commit properly creates the symlinks from the files passed to
the `download()` function so that this problem no longer happens.