[now-node-bridge] Refactor to handle `unhandledRejection` events
Currently, if you have an async handler function that throws an error,
an `unhandledRejection` event occurs and the HTTP request never
completes and hangs until the browser or lambda invoke timeout is reached.

The commit makes it such that the lambda invoke returns the error
properly so that the HTTP request will return a 502 quickly.
This involves using the "deferred promise" pattern to reject the lambda
promise during the `unhandledRejection` event, and some logic to tear
down the connected socket.

The test case is a standalone node script because jest does some magic
behind the scenes to fail the test when an `unhandledRejection` event
happens, even though we have our own listener that does handle it.