[now-static-build] Use `yarn` when invoking `now-dev` script
`yarn` used to be used, but was switched over to `npm` in #636.

The problem with `npm` is that when the `ignoreScripts` config option
is enabled in the user's `.npmrc` file, then `npm run …` does not get
invoked. This makes sense for `npm i`, but doesn't really make sense if
you're explicitly _trying_ to run a package script via `npm run`, and in
fact breaks `now dev`.

`yarn` has a similar option to ignore scripts, however it has the more
understandable behavior that when you do `yarn run` then the script is
actually run.

There isn't a concern about if the user does not have `yarn` installed
since `now dev` bundles its own copy of yarn and uses that version.